TST: updates to BaseOptimiserEngine unit testing
