+++ force_bdss/mco/optimizer_engines/tests/test_base_optimizer_engine.py
@@ -6,7 +6,7 @@ from unittest import TestCase
 from force_bdss.api import KPISpecification, RangedMCOParameterFactory
 from force_bdss.tests.dummy_classes.mco import DummyMCOFactory
 from force_bdss.tests.dummy_classes.optimizer_engine import (
-    EmptyOptimizerEngine,
+    DummyOptimizerEngine,
 )
 from force_bdss.tests.probe_classes.workflow_file import ProbeWorkflowFile
 from force_bdss.tests import fixtures
@@ -21,24 +21,23 @@ class TestBaseOptimizerEngine(TestCase):
         self.workflow = workflow_file.workflow
 
         self.parameters = [1, 1, 1, 1]
-        self.optimizer_engine = EmptyOptimizerEngine(
+        self.optimizer_engine = DummyOptimizerEngine(
             single_point_evaluator=self.workflow
         )
 
     def test_initialize(self):
-        self.assertIs(
-            self.workflow, self.optimizer_engine.single_point_evaluator
-        )
         self.assertListEqual([], self.optimizer_engine.parameters)
         self.assertListEqual([], self.optimizer_engine.kpis)
 
     def test_base_methods(self):
-        self.assertListEqual([0.0], self.optimizer_engine.optimize())
-
         point = [1.0]
-        self.assertListEqual(
-            self.workflow.evaluate(point), self.optimizer_engine._score(point)
+        kpi_values = self.workflow.evaluate(point)
+        score = self.optimizer_engine._score(point)
+
+        self.assertDictEqual(
+            {(1.0,): kpi_values}, self.optimizer_engine._kpi_cache
         )
+        self.assertEqual(0, score.size)
 
     def test_parameter_bounds(self):
         self.optimizer_engine.parameters = [
@@ -55,6 +54,18 @@ class TestBaseOptimizerEngine(TestCase):
         self.assertListEqual(
             self.optimizer_engine.parameter_bounds,
             [(0.0, 1.0)] * len(self.parameters),
+        )
+
+    def test_kpi_bounds(self):
+        self.optimizer_engine.kpis = [
+            KPISpecification(
+                lower_bound=0.0, upper_bound=1.0
+            )
+        ]
+
+        self.assertListEqual(
+            [(0.0, 1.0)],
+            self.optimizer_engine.kpi_bounds,
         )
 
     def test__minimization_score(self):
